Day 11 - Solution 1 + 2
This one was simpler as I first thought: I decided to use a Coordinate Set to store the galaxies (not the empty spaces),
which turned out to be a very good decision. The most complex part was to find empty cols/rows and insert empty ones.
The distance calculation was piece of cake: the 'shortest path' were simply all Manhattan Distances between pairs.