Day 16 - Solutions 1 + 2
Straight-forward walking through a 'maze' of sorts. I decided to use a breadth-first algorithm using a queue to add up the next moves.

The 2nd solution could of course be optimized a bit, now it runs for about 3 seconds. I think I will try a Threaded version later in the day.